Finish the Drum Kit.
Add all drum kit buttons in index.html and then style it.
Use a function that takes in the button pressed to animate the
button by adding and removing the .pressed css style. Use a
function to make a switch case for every drum sound to play.
Use a for loop to add an event listener for clicks to all buttoons
and trigger sounds and button animation. Finally add an event listener
for key presses to do the same.